+++ addon/src/main/resources/VAADIN/mediumeditor/js/medium-editor-connector.js
@@ -1,6 +1,8 @@
 window.com_byteowls_vaadin_mediumeditor_MediumEditor = function() {
   // see the javadoc of com.vaadin.ui.AbstractJavaScriptComponent for all functions on this.
+  var THEME_ID = "medium-editor-theme-id";
   var mediumEditor;
+  var mediumThemeLink;
   var e = this.getElement();
   // Please note that in JavaScript, this is not necessarily defined inside callback functions and it might therefore be necessary to assign the reference to a separate variable
   var self = this;
@@ -14,6 +16,21 @@ window.com_byteowls_vaadin_mediumeditor_MediumEditor = function() {
     loggingEnabled = state.loggingEnabled;
     focusOutlineEnabled = state.focusOutlineEnabled;
     readOnly = state.readOnly;
+
+    // #1 theme
+    if (typeof mediumThemeLink === 'undefined') {
+      // check if a theme css is already loaded
+      mediumThemeLink = document.getElementById(THEME_ID);
+      if (mediumThemeLink == null) {
+        mediumThemeLink = document.createElement("link");
+        mediumThemeLink.setAttribute("rel", "stylesheet");
+        mediumThemeLink.setAttribute("type", "text/css");
+        mediumThemeLink.setAttribute("id", THEME_ID);
+        document.getElementsByTagName("head")[0].appendChild(mediumThemeLink);
+      }
+    }
+    var url = this.translateVaadinUri("vaadin://mediumeditor/css/"+state.theme.toLowerCase()+".min.css");
+    mediumThemeLink.setAttribute("href", url);
 
     if (loggingEnabled) {
       console.log("medium-editor: setting value to\n" + state.content);